Naive conversion to use BPF for filters.
This just uses libpcap, when it should include the JIT instead, and
probably brilter_pass should be hand-inlined.  And it'd be nice to be
able to pull the filters from a file.  And maybe even support a more
complex set of matching criteria and decisionmaking.  Using Lua to
generate the BPF might make sense, along with additional control
structures.